Refactored the structure of the plugin and renamed "dansplugins" to "preponderous"
