--- src/main/java/dansplugins/ponder/specification/IPonderAPI_Integrator.java
@@ -1,8 +1,0 @@
-package dansplugins.ponder.specification;
-
-import dansplugins.ponder.implementation.PonderAPI;
-
-public interface IPonderAPI_Integrator {
-    boolean isPonderPresent();
-    PonderAPI getAPI();
-}
