--- src/main/java/dansplugins/ponder/specification/commands/ICommand.java
@@ -1,8 +1,0 @@
-package dansplugins.ponder.specification.commands;
-
-import org.bukkit.command.CommandSender;
-
-public interface ICommand {
-    boolean execute(CommandSender sender);
-    boolean execute(CommandSender sender, String[] args);
-}
