--- src/main/java/dansplugins/ponder/specification/services/ICommandService.java
@@ -1,7 +1,0 @@
-package dansplugins.ponder.specification.services;
-
-import org.bukkit.command.CommandSender;
-
-public interface ICommandService {
-    boolean interpretCommand(CommandSender sender, String label, String[] args);
-}
