--- src/main/java/dansplugins/ponder/specification/services/IStorageService.java
@@ -1,13 +1,0 @@
-package dansplugins.ponder.specification.services;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public interface IStorageService {
-    void save();
-    void load();
-    void writeOutFiles(List<Map<String, String>> saveData, String fileName);
-    ArrayList<HashMap<String, String>> loadDataFromFilename(String filename);
-}
