--- src/main/java/dansplugins/ponder/specification/toolbox/IToolbox.java
@@ -1,12 +1,0 @@
-package dansplugins.ponder.specification.toolbox;
-
-import dansplugins.ponder.specification.toolbox.tools.*;
-
-public interface IToolbox {
-    IArgumentParser getArgumentParser();
-    IEventHandlerRegistry getEventHandlerRegistry();
-    ILogger getLogger();
-    IPermissionChecker getPermissionChecker();
-    IScheduler getScheduler();
-    IUUIDChecker getUUIDChecker();
-}
