--- src/main/java/dansplugins/ponder/specification/toolbox/tools/IArgumentParser.java
@@ -1,8 +1,0 @@
-package dansplugins.ponder.specification.toolbox.tools;
-
-import java.util.ArrayList;
-
-public interface IArgumentParser {
-    String[] dropFirstArgument(String[] args);
-    ArrayList<String> getArgumentsInsideDoubleQuotes(String[] args);
-}
