--- src/main/java/dansplugins/ponder/specification/toolbox/tools/ILogger.java
@@ -1,6 +1,0 @@
-package dansplugins.ponder.specification.toolbox.tools;
-
-public interface ILogger {
-    void log(boolean debug, String message);
-    void print(String message);
-}
