--- src/main/java/dansplugins/ponder/specification/toolbox/tools/IPermissionChecker.java
@@ -1,7 +1,0 @@
-package dansplugins.ponder.specification.toolbox.tools;
-
-import org.bukkit.command.CommandSender;
-
-public interface IPermissionChecker {
-    boolean checkPermission(CommandSender sender, String permission);
-}
