--- src/main/java/dansplugins/ponder/specification/toolbox/tools/IScheduler.java
@@ -1,5 +1,0 @@
-package dansplugins.ponder.specification.toolbox.tools;
-
-public interface IScheduler {
-    void scheduleAutosave();
-}
