--- src/main/java/dansplugins/ponder/specification/toolbox/tools/IUUIDChecker.java
@@ -1,9 +1,0 @@
-package dansplugins.ponder.specification.toolbox.tools;
-
-import java.util.UUID;
-
-public interface IUUIDChecker {
-    String findPlayerNameBasedOnUUID(UUID playerUUID);
-
-    UUID findUUIDBasedOnPlayerName(String playerName);
-}
