+++ src/main/java/preponderous/ponder/commands/specification/ICommand.java
@@ -1,0 +1,8 @@
+package preponderous.ponder.commands.specification;
+
+import org.bukkit.command.CommandSender;
+
+public interface ICommand {
+    boolean execute(CommandSender sender);
+    boolean execute(CommandSender sender, String[] args);
+}
