+++ src/main/java/preponderous/ponder/services/specification/ICommandService.java
@@ -1,0 +1,7 @@
+package preponderous.ponder.services.specification;
+
+import org.bukkit.command.CommandSender;
+
+public interface ICommandService {
+    boolean interpretCommand(CommandSender sender, String label, String[] args);
+}
