+++ src/main/java/preponderous/ponder/services/specification/IConfigService.java
@@ -1,0 +1,16 @@
+package preponderous.ponder.services.specification;
+
+import org.bukkit.command.CommandSender;
+import org.bukkit.configuration.file.FileConfiguration;
+
+public interface IConfigService {
+    void saveMissingConfigDefaultsIfNotPresent();
+    void setConfigOption(String option, String value, CommandSender sender);
+    void sendConfigList(CommandSender sender);
+    boolean hasBeenAltered();
+    FileConfiguration getConfig();
+    int getInt(String option);
+    boolean getBoolean(String option);
+    double getDouble(String option);
+    String getString(String option);
+}
