+++ src/main/java/preponderous/ponder/specification/IPonderAPI_Integrator.java
@@ -1,0 +1,8 @@
+package preponderous.ponder.specification;
+
+import preponderous.ponder.PonderAPI;
+
+public interface IPonderAPI_Integrator {
+    boolean isPonderPresent();
+    PonderAPI getAPI();
+}
