+++ src/main/java/preponderous/ponder/toolbox/tools/specification/IArgumentParser.java
@@ -1,0 +1,8 @@
+package preponderous.ponder.toolbox.tools.specification;
+
+import java.util.ArrayList;
+
+public interface IArgumentParser {
+    String[] dropFirstArgument(String[] args);
+    ArrayList<String> getArgumentsInsideDoubleQuotes(String[] args);
+}
