+++ src/main/java/preponderous/ponder/toolbox/tools/specification/ILogger.java
@@ -1,0 +1,6 @@
+package preponderous.ponder.toolbox.tools.specification;
+
+public interface ILogger {
+    void log(boolean debug, String message);
+    void print(String message);
+}
