+++ src/main/java/preponderous/ponder/toolbox/tools/specification/IPermissionChecker.java
@@ -1,0 +1,7 @@
+package preponderous.ponder.toolbox.tools.specification;
+
+import org.bukkit.command.CommandSender;
+
+public interface IPermissionChecker {
+    boolean checkPermission(CommandSender sender, String permission);
+}
