+++ src/main/java/preponderous/ponder/toolbox/tools/specification/IScheduler.java
@@ -1,0 +1,5 @@
+package preponderous.ponder.toolbox.tools.specification;
+
+public interface IScheduler {
+    void scheduleAutosave();
+}
