+++ src/main/java/preponderous/ponder/toolbox/tools/specification/IUUIDChecker.java
@@ -1,0 +1,9 @@
+package preponderous.ponder.toolbox.tools.specification;
+
+import java.util.UUID;
+
+public interface IUUIDChecker {
+    String findPlayerNameBasedOnUUID(UUID playerUUID);
+
+    UUID findUUIDBasedOnPlayerName(String playerName);
+}
